requirements.txt
python-dotenv
numpy
pandas
pymongo
scikit-learn
matplotlib
seaborn
lightgbm
statsmodels
setuptools
pymongo
scipy
dill
pyyaml  
certifi
xgboost
lightgbm
catboost
mlflow
##pymongo[srv]==3.11
##-e .
